API/requirements.txt
# Flask
flask==1.1.2
flask-sqlalchemy==2.4.4
flask-script==2.0.6
flask-restful==0.3.8
flask-bcrypt==0.7.1
flask-cors==3.0.9

# Base datos
py-postgresql==1.2.2
psycopg2-binary==2.8.6
marshmallow==3.8.0

# Encriptado
pyjwt==1.7.1
bcrypt==3.2.0

# Enviroment
python-dotenv==0.14.0
flask-dotenv==0.1.2

# Utils
flask-mail==0.9.0
pip==20.2.4

gunicorn==20.1.0